% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{run_app}
\alias{run_app}
\title{Run SASPACer Shiny App}
\usage{
run_app()
}
\value{
No return value, launches a Shiny app.
}
\description{
This function launches SASPACer application
}
\examples{
# run_app()
}
