% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SASPACerShiny-package.R
\docType{package}
\name{SASPACerShiny}
\alias{SASPACerShiny-package}
\alias{SASPACerShiny}
\title{SASPACerShiny: Shiny application version of SASPACer}
\description{
A Shiny-based graphical user interface for creating SAS packages.
}
\author{
\strong{Maintainer}: Ryo Nakaya \email{nakaya.ryou@gmail.com}

}
