% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{run_app}
\alias{run_app}
\title{Run SASPACer Shiny App}
\usage{
run_app()
}
\value{
No return value, launches a Shiny app.
}
\description{
This function starts a Shiny app that allows users to create SAS packages source
folders/files.
}
\details{
This function launches SASPACer shiny application. SASPACer is a tool originally made in SAS.
The SAS package SASPACer is excel-based support tool for creating SAS packages, while
SASPACer shiny app is atool to support creating SAS packages via user-friendly GUI.
}
\examples{
# run_app()

}
